refactor(settings): each section item in own component file
