+++ packages/suite/src/views/settings/debug/InvityApi.tsx
@@ -1,0 +1,61 @@
+import React from 'react';
+import styled from 'styled-components';
+
+import { ActionColumn, ActionSelect, SectionItem, TextColumn } from '@suite-components/Settings';
+import * as suiteActions from '@suite-actions/suiteActions';
+import { useSelector, useActions } from '@suite-hooks';
+import invityAPI from '@suite-services/invityAPI';
+
+const StyledActionSelect = styled(ActionSelect)`
+    min-width: 250px;
+`;
+
+const invityApiServerOptions = [
+    {
+        label: invityAPI.productionAPIServer,
+        value: invityAPI.productionAPIServer,
+    },
+    {
+        label: invityAPI.stagingAPIServer,
+        value: invityAPI.stagingAPIServer,
+    },
+    {
+        label: invityAPI.localhostAPIServer,
+        value: invityAPI.localhostAPIServer,
+    },
+];
+
+export const InvityApi = () => {
+    const { setDebugMode } = useActions({
+        setDebugMode: suiteActions.setDebugMode,
+    });
+    const { debug } = useSelector(state => ({
+        debug: state.suite.settings.debug,
+    }));
+
+    const selectedInvityApiServer =
+        invityApiServerOptions.find(s => s.value === debug.invityAPIUrl) ||
+        invityApiServerOptions[0];
+
+    return (
+        <SectionItem>
+            <TextColumn
+                title="API server"
+                description="Set the server url for buy and exchange features"
+            />
+            <ActionColumn>
+                <StyledActionSelect
+                    noTopLabel
+                    onChange={(item: { value: string; label: string }) => {
+                        setDebugMode({
+                            invityAPIUrl: item.value,
+                        });
+                        invityAPI.setInvityAPIServer(item.value);
+                    }}
+                    value={selectedInvityApiServer}
+                    options={invityApiServerOptions}
+                />
+            </ActionColumn>
+        </SectionItem>
+    );
+};
