+++ packages/suite/src/views/settings/debug/SettingsDebug.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+
+import { SettingsLayout } from '@settings-components';
+import { Section } from '@suite-components/Settings';
+import { isWeb } from '@suite-utils/env';
+
+import { TranslationMode } from './TranslationMode';
+import { GithubIssue } from './GithubIssue';
+import { WipeData } from './WipeData';
+import { InvityApi } from './InvityApi';
+import { DarkModePalette } from './DarkModePalette';
+
+const SettingsDebug = () => (
+    <SettingsLayout>
+        {isWeb() && (
+            <Section title="Localization">
+                <TranslationMode />
+            </Section>
+        )}
+        <Section title="Debug">
+            <GithubIssue />
+            {!isWeb() && <WipeData />}
+        </Section>
+        <Section title="Invity">
+            <InvityApi />
+        </Section>
+
+        <DarkModePalette />
+    </SettingsLayout>
+);
+
+export default SettingsDebug;
