+++ packages/suite/src/views/settings/device/Passphrase.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { Translation } from '@suite-components';
 import { ActionColumn, SectionItem, TextColumn } from '@suite-components/Settings';
 import { Switch } from '@trezor/components';
@@ -6,11 +7,11 @@ import { useDevice, useAnalytics, useActions } from '@suite-hooks';
 import * as deviceSettingsActions from '@settings-actions/deviceSettingsActions';
 import { PASSPHRASE_URL } from '@suite-constants/urls';
 
-interface Props {
+interface PassphraseProps {
     isDeviceLocked: boolean;
 }
 
-const Passphrase = ({ isDeviceLocked }: Props) => {
+export const Passphrase = ({ isDeviceLocked }: PassphraseProps) => {
     const { device } = useDevice();
     const { applySettings } = useActions({
         applySettings: deviceSettingsActions.applySettings,
@@ -46,4 +47,3 @@ const Passphrase = ({ isDeviceLocked }: Props) => {
         </SectionItem>
     );
 };
-export default Passphrase;
