+++ packages/suite/src/views/settings/general/Analytics.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
+
 import { Switch } from '@trezor/components';
 import { Translation } from '@suite-components/Translation';
 import { SectionItem, ActionColumn, TextColumn } from '@suite-components/Settings';
@@ -9,15 +10,14 @@ const PositionedSwitch = styled.div`
     align-self: center;
 `;
 
-const Analytics = () => {
+export const Analytics = () => {
     const { enable, dispose, enabled } = useAnalytics();
+
     return (
         <SectionItem>
             <TextColumn
                 title={<Translation id="TR_ALLOW_ANALYTICS" />}
                 description={<Translation id="TR_ALLOW_ANALYTICS_DESCRIPTION" />}
-                // todo: disabled until we get where to redirect
-                // learnMore="todo some link"
             />
             <ActionColumn>
                 <PositionedSwitch>
@@ -36,5 +36,3 @@ const Analytics = () => {
         </SectionItem>
     );
 };
-
-export default Analytics;
